alpha version

+++ src/main/model/Heuristic/Heuristic.scala
@@ -3,5 +3,5 @@ package main.model.Heuristic
 import main.model.MyState
 
 abstract class Heuristic {
-  def calc(actual : MyState, future : MyState, player : Boolean) : Double
+  def calc(actual : MyState, future : MyState, player : Boolean) : Float
 }
